requirements.txt
et_xmlfile==2.0.0
lxml==6.0.2
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pillow==12.0.0
pypdf==6.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
xlrd==2.0.2
xlsxwriter==3.2.9
pytest
google-adk
